Skips bandit check on dom0 updater tmpfile
As explained in the comment, we accept the risk of a hardcoded tempfile
in dom0, because it's a single-user system, isolated from the other VM
components. At a later date, we'll likely port the check-migration logic
to pure python, but unfortunately 'import qubesadmin' would break the
mocked CI tests for the launcher and updater.