Adjusts GUI updater run order
Ensuring that we're able to ship unattended migrations in the future.
It's important that *first* we update the dom0 RPM, if required, then
re-apply the dom0 state. After that, we'll check whether a migration is
requested (that's the clincher), then either run a full-state apply or
upgrade packages within each TemplateVM, depending on the check.

Tinkered with the progress bars accordingly. It's not ideal, since the
full-state run blocks the progress bar.