Adds exit 0 to sd-log rc.local script
Since the /etc/sd-rsyslog.conf file is not yet in place on sd-log,
trying to start the "securedrop-log" service will fail. Although the
rc.local script doesn't exit on error, it will return the final error
code of a command run within it. Since the securedrop-log task is last,
it'll exit non-zero on fresh installs.