Adds migration flag for mailcap-hardening
The updater must perform a full state run in order to apply the mailcap
hardening inside AppVM private volumes. Create the necessary
run-everything flag via RPM postinst. The updater logic will
automatically clean up the flags are it runs. We'll need to remove this
logic from postinst in the subsequent release of
'securedrop-workstation-dom0-config', so it doesn't recreate the flag.