[ci] Update containers for Debian PyQt
In Qubes 4.0, dom0 is based on F25, which uses python3-pyqt4.
It'd be a pain to maintain our own F25 container to pull in the
Qubes-maintained backport of python3-pyqt4, so instead we'll stick with
Debian Buster (oldstable) which has both python3-pyqt4 & python3-pyqt5
available for installation.

Developers should use the appropriate system package, installed via apt,
to choose the environment they want to develop on.

While making these changes, I updated the dev-dependencies across the
board, so that identical versions of e.g. black where used in all
contexts.