Allow qubes.VMRootShell from disp-mgmt-sd-*-templ
The R4.0 equivalent of these lines were temporarily injected during the
updating process. This isn't true for all templates however, and I've
not found what triggers this to be required, and if R4.1 removed the
logic or if there's an equivalent that could be used instead.

Should probably be discussed as to what implications this has, or looked
into how it can be avoided.