Separate RPC policies 4.0/4.1
The Qubes RPC policy ststem uses different formats
for Qubes 4.0 & 4.1.  Fortunately, there's decent
backwards-compatibility in 4.1, but it's not 100%
out of the box. In order to minimize changes for
adding 4.1 compat for SDW, let's separate into:

  * same filepath and content for both versions
  * 4.0-specific
  * 4.1-specific

This change does so, but lacks testing coverage:
the RPC tests only work on 4.0. Changes TK.