Harden reproducibility
%install scriptlets will have certain macros expand automatically - one
of that is %__spec_install_post which will run
`/usr/lib/rpm/redhat/brp-mangle-shebangs` which can affect modification
dates for the folders containing scripts which have gotten their shebang
mangled.

So we work around that by running the whole macro manually before we
enforce modification times for reproducibility, and subsequently
redefining the macro so it doesn't run unnecessarily.