Run flake8 normally and re-enable in CI
The lint-all script was buggy (the mime is now text/x-script.python), so
it was actually excluding files that would've normally been detected.
Instead explicitly list the 4 extension-less Python files in the flake8
configuration.

To prevent this from being missed again, add a `make lint` target that
both CI and developers can use. Additional tools like isort and mypy can
be added to this as well.

Fixes #855.