Configure mypy for type checking
This is slightly different than other repositories because all the
code is in individual scripts rather than a typical Python package.

Also we don't have type stubs for qubesadmin so we need to ignore
missing import errors.

All of the launcher/ code is ignored since it's moving to
securedrop-updater and will be removed from here soon.

Fixes #854.