Make our VMs internal, hiding their applications from the menu
We want our users to interact with the system through the client, which
is started via the updater, rather than directly with a VM which would
be the normal Qubes way. This should make the menu a bit less noisy as
well.

The omission of sd-whonix is intentional, as it contains a couple of Tor
related GUI tools that users might need access to and have no other way
to get to.