Make our VMs internal for prod setups, hiding their applications from the menu.
We want our users to interact with the system through the client, which
is started via the updater, rather than directly with a VM which would
be the normal Qubes way. This should make the menu a bit less noisy as
well.

The omission of sd-whonix and sd-devices is intentional.
sd-whonix contains a couple of Tor related GUI tools that users might
need access to and have no other way to get to. sd-devices is sometimes
accessed to manually unlock VeraCrypt drives and interact with USB
peripherals.