Add versioned migrations
This is a test-balloon for `securedrop-workstation`, where we need
complex migrations that don't rely on having to install every package
incrementally, but allows for upgrading to the newest package directly.

When the package is upgraded, a version number of the previous installed
version is read and migrations that form the delta between the target
version are run in succession.

To allow the use of this pattern immediately, the version number is also
checked on installation, letting us deploy the package with a`%post`
scriptlet that ensures an "initial migration" is run.

`%post` scriptlets are always expected to run successfully, and even if
scriplets fail, the package installation isn't marked as failed. Hence,
later down the line we'll need logic in the updater itself to check for
the expected version, and respond appropriately if a migration failed.