Force a full migration for Fedora 37 template upgrade
The files needed for USB auto-attachment are only copied into
sd-fedora-37-dvm during a full `sdw-admin --apply` run, which doesn't
happen by default because it's slow.

Set a flag file to tell the updater to do a full migration. This
is basically the same as was used in d4a8d1ba60f7.

Refs #866.