Don't verify a specific kernel version
We are now shipping kernels more frequently, so it's not useful to
assert an exact kernel version is being used, continually needing
to be manually updated.

As of <https://github.com/freedomofpress/securedrop/commit/9cd5e8d107a1997ffcb9e061113932911ab20bfe>
in SecureDrop server, we only verify that the running kernel is using
the correct grsec flavor, which is what I've updated this test to match.

Fixes #872.