Just set `Version` in RPM spec
This is mostly a copy of @eaon's change in securedrop-updater[1], with
one key change of normalizing the version down to a PEP 440 compatible
format.

Since we use setuptools's sdist to create a tarball and then expect
rpmbuild to be able to find it, we need to use a common version format
across both setuptools and RPM. Encoding the RC version in the `Release`
field has always been wrong, since that's used for changes to packaging,
not when the upstream (i.e. sdist tarball) changes.

This also means we no longer need to update `Source` in the release
branch to point at the tarball, since it can be inferred correctly from
the existing `%{name}-%{version}.tar.gz` macros.

Leave a TODO to switch to rpmdev-bumpspec which is the RPM `dch`
equivalent to bump the version and insert a changelog entry.

[1]
https://github.com/freedomofpress/securedrop-updater/commit/42a55c7f10546710e99fa73ae291a69b04dc4eeb

Co-authored-by: Michael Z <michael@freedom.press>