Don't install dev deps in the RPM build container
The RPM build container needs to be highly trustworthy, so it should
only contain things from Fedora itself and any audited code. We don't
audit development dependencies, so we shouldn't be installing them into
the build container.

We split the container used by `./scripts/container.sh` into two, a
base, build container and then a container layered on top with dev
dependencies.

Functionally this should be a no-op since none of the dependencies are
used at build time but it cuts down on the risk of malicious code
injection.

Fixes #921.