Create base template by cloning debian-12-minimal
Instead of shipping our own template, we now pull the debian-12-minimal template,
clone it, install some base packages, then clone it again into the small and
large variants.

Take the opportunity to switch the kernel to use PVH instead of HVM.

While people should only be using fresh Qubes installs for SDW, a check is added
to ensure there are no other VMs using the debian-12-minimal template, since we
need it to be pristine.