Temporary: Shutdown our VMs after provisioning
While we do recommend a reboot, just preemptively shutdown all of
our VMs after provisioning to ensure anything that follows (e.g. CI
runs) will be from a cleaner state.

We tried to implement this in workstation-ci itself (<https://github.com/freedomofpress/securedrop-workstation-ci/pull/44>),
but it isn't working as we desired, so we'll do this for now and revert
once CI is fixed.