Make _validate_apt_sources more robust against salt's non-determinism
salt parses the sources.list line we give it and non-deterministically
re-orders the components, so sometimes our direct assertion fails with
the components out of order.

We want to switch to deb822 format .sources file (<https://github.com/freedomofpress/securedrop-client/pull/1952>),
so for now just assert against both orderings of the URL as a
short-term fix.

Fixes #982.