Switch to poetry, split up `make install-deps`
Qubes 4.2's dom0 ships with Python 3.11, which is much easier to deal
with than 4.1's 3.8, since it matches with Debian bookworm. As a result,
we can expect most developers to be able to natively use 3.11 and not
need a container wrapper.

It's not straightforward to just swap out pip for poetry here since
poetry sets up a virtualenv per-user instead of globally installing.

At the same time, the `make install-deps` target has grown so that it
encompases both build and test dependencies and is ripe for splitting.
It also had some duplication in that build dependencies were specified
in both the Makefile and specfile.

So we can address everything at once. There are now three effective sets
of dependency groupings:

1) Python lint/test dependencies, specified in pyproject.toml and
installed using poetry.
2) Build dependencies, specified in the `.spec` file and installed via
`make build-deps`.
3) Test dependencies, specified in the Makefile and installed via `make
test-deps`.

The corresponding make targets, container.sh and CI manifests have been
updated to handle this. The `install-deps` target was removed so it
isn't accidentally used anymore.

The launcher dependencies were identical except that virtualenv also
needed PyQt5. We already had a root `make test-launcher`, so fix that up
and get rid of separate launcher dependencies and Makefile.

Fixes #992.
Fixes #947.