Clean up shebangs so they don't need mangling
rpm has some support for mangling shebangs, but it is sometimes flaky
under reprotest (#1010). We should understand that better, but we can
also avoid the need for inconsistent mangling by fixing them up in
source itself. All of the Python files are executed by the system
interpreter and not a virtualenv, so they shouldn't go through `env`
first.

Modern Fedora and Debian Bookworm have had /usr merged, so /bin is a
symlink to /usr/bin, so we can specify the full path to /usr/bin/bash in
shebangs.

Fixes #1010.