test(test_v3_auth_private_file): accept ".onion" suffix
Despite what's documented[1], Tor seems happy to configure client
authentication for an onion address ending in ".onion".  That is, I'm
able to break "app-journalist.auth_private" by mangling it in other
ways, but it works both with and without ".onion".

If need be, we can always change the semantics of the
"/vm-config/SD_HIDSERV_HOSTNAME" QubesDB key to distinguish between
versions with and without the ".onion" suffix; or remove it and let
applications add it where they need it.

[1]: https://gitlab.torproject.org/tpo/core/tor/-/blob/7a5d94bcf842299534b667433424ac7a1133d371/doc/man/tor.1.txt#L3659