Ensure and verify files in built RPM have their mtime clamped
The mtime of all files in the RPM is supposed to be clamped to the
changelog, which is also set as the build date. "clamp" means that the
file's mtime will be no greater than SOURCE_DATE_EPOCH, not that it will
be fixed to it (like in Debian packages). So if you have a file with an
old mtime, it will not be adjusted by rpm, leading to a non-reproducible
build.

We now touch every file to a date in the future so rpm can clamp them
back to the changelog date, which hopefully will not also be in the
future.

Then a new script verifies the RPM has all of its files set to that
clamped time, and if not, causes the build to fail.

While we're at it, make builds more verbose by removing `--quiet` so we
can actually see the details of what RPM is doing under the hood.

This commit is compatible with Python 3.8 so it can be backported for
Qubes 4.1 / 0.11.1.

Fixes #1110.

(cherry picked from commit e2fa8c7627572ca8e3b12bcd61c981adce1c8d37)