Replace salt updater
Replace salt updater with qubes-vm-update in sd-sys-vms. Suppress
output since it is likely to be verbose due to this being the first
update the template receives since it was built.

Unlike the salt updater, in with qubes-vm-update if a target qube
does not exist, the updater will fail. This is what caused #1165
to go undetected.