Remove perl from tests
There's no reason to use perl here, given that we can do the same
manipulations in Python itself in a way that's more maintainable
and approachable.

I took the opportunity to consolidate the two mimeapps tests in
test_app into one, but decided to just copy it over to test_proxy_vm
instead of moving it to the base test to avoid too much refactoring
here besides the deperlification.

Refs #988.