Always set a VM's internalness
To ensure we correctly flip the status of a VM's internal feature or
not, ensure the property is always being set, but conditional the value
on the environment.

And then explicitly set sd-app and sd-gpg as non-internal, so that salt
will update them from when they used to be internal (see e754b40e487).

Refs #1214