Add tests to verify prod signing key
Based on the keyring tests in securedrop-client, add a test for the PGP
key embedded in apt_freedom_press.sources and the key for dom0.

It verifies only one key is present, it has the fingerprint we expect,
it has a 2027 expiry, and the expiry is more than 6 months from now.