Configure sd-log at the same time as other VMs
Historically sd-log (and its template) was being configured earlier
than the remaining qubes so that logs would be captured early in the
install process. However, is some cases restarting sd-log hangs due
the already-configured templates trying to send logs to it, which
starts sd-log, while it is getting shut down (#1247).

The current solution to avoid this race-condition and adhering more
to the salt convention of not trying to force order outside of salt,
sd-log will no longer be setup early.

After a full system restart sd-log should be capturing logs as it
normally would.