Salt: reacting to fedora install by using 'onchanges'
Previously, the 'watch' salt requisite was used to detect if
the template installation had had changes (i.e. if a new fedora
template had been installed). This was not working as intended
and with the replacement of 'cmd.wait' with 'qvm.template_installed'
it would be impossible due to a missing 'mod_watch' in the Qubes
salt state modules, according to the docs:

> The watch requisite only works if the state that is watching has
> a mod_watch function written. If the watching state where the
> watch is set does not have a mod_watch function (like pkg), then
> the listed states will behave only as if they were under a
> require statement.
>
> -- https://docs.saltproject.io/salt/user-guide/en/latest/topics/requisites.html#the-watch-requisite