Remove hardcoded whonix version from python code
Qubes's salt formulas are the cannonical location for getting the
Whonix version installed / to be installed. Specifically in
qvm/whonix.jinja, a jinja-template file.

Therefore python code needs a way to extract it. The best way
found is to use salt to load the file as a jinja map and extract
the `whonix_version` variable.

  sudo qubesctl jinja.load_map qvm/whonix.jinja whonix_version --out newline_values_only
                       |          |                    |                      |
                       |          |              variable name                |
                       |          +--> jinja file (relative to file_roots)    |
       extracts a value from jinja file                                       |
                                                                   just print the value without
                                                                   stating which minion it ran on
                                                                   (in this case it be "local")

(cherry picked from commit 3308118cde69646490f13366766589ec0ebc3e6f)