Tests: have static whonix version instead
It should be acceptable to have tests with a hardcoded whonix
version. When whonix 18 comes around the tests may break and we
just need to update them. Otherwise, we may risk there being
some problem obtaining the whonix version and it getting missed
because we are comparing that value with itself.