Tests: distinguish provisioning from configuration
With the use of fixtures, a test structure change can help
better clarify the scope of fixtures: /conftest.py is for shared
fixtures, /configure/conftest.py for configuration fixtures and
the same for dom0 tests.

Since the project being tested is not a typical python project,
the suggestion to keep the test structure to the code structure
does not really apply as far as provision and configuration tests
are concerned.