Trigger OpenQA from GitHub Actions
A GHA job will trigger our jobs in OpenQA and then monitor them until
they finish.

The OpenQA job will only be started if it is a direct push that is to a
branch that isn't `main` nor part of the GH merge queue. This is to
avoid duplicate jobs while it's still quite slow; eventually this should
be a pre-merge check.

Because this is a slow job running on Qubes hardware, it's trivial to
skip running it if the latest commit message mentions the string "skip"
and "openqa" separated by a single space, case insensitive. This is
what the `check-openqa` job does and is based off the "skip" "ci"
keywords that GHA itself supports.

When the GHA job is cancelled, we'll also send off requests to cancel
the openQA jobs in a best-effort manner.

Finally we include links to the openQA logs at the bottom just to make
it easy for humans to hunt down failures.

This work was initially based off of
https://github.com/os-autoinst/os-autoinst-distri-example/blob/main/.github/workflows/openqa.yml
and deeplow's example in #1319.

Note that the `openqa-cli monitor` command is only available in trixie.

Fixes #1319.