Add workaround for SELinux fedora breakage
Fixes #1369. Fedora update was failing in certain templates. Out of
abundance of caution, we're disabling SELinux (an acceptable risk) on
a temporary basis. This allows the templates to successfully update
until the original issue is fixed. Otherwise fedora-based VMs won't
boot. (to be tested on openqa)

Backported from commit 9e088d