Remove Xvfb dep when running launcher tests in dom0
A virtual framebuffer is critical when running tests in a CI environment
where no display is available. But with OpenQA we can now run these in a
real-like system. Removing this dependency is one less thing we need to
install in dom0 for its tests.

NOTE: windows may flash as a result of the virtual frame buffer removal.