Reenable SELinux in Fedora
Removal of salt state should effectively leave SELinux enabled, since
this is the default on Fedora templates on Qubes and we're doing this
while ensuring a newly released template is installed.

Reverts a workaround added in commit 9e088d for the hotfix release 1.2.1
workstation but this was information we did not have at the time of the
release.

Fixes #1375