Skip individual launcher tests instead of per file
test_signing_key.py and test_sources.py previously had pytest.skip
before running imports not available in dom0. Even though this allowed
for the successful execution in dom0, the issue was that the JunitXML
report got an empty classname:

  <testcase classname="" name="launcher.tests.test_sources" time="0.000">
    <skipped message="collection skipped">
      'Skipped: not running due to unavailable dom0 python modules')
    </skipped>
  </testcase>

This caused issues with the convert_junit.py script [1], which make it
compatible with openqa. Instead of making the script compatible, a
simpler fix was to skip individual tests and skip the problematic dom0
imports. After this, the generated reports have classname:

<testcase classname="launcher.tests.test_signing_key" name="test_apt_sources" time="0.000">
  <skipped type="pytest.skip" message="Test cannot be run in dom0">
     Test cannot be run in dom0 due to missing modules
  </skipped>
</testcase>

[1]: https://github.com/QubesOS/openqa-tests-qubesos/blob/46e01db/extra-files/convert_junit.py