Remove "sd-whonix" from salt, update and tests
Removes sd-whonix from salt code, but adding steps to remove remove
sd-whonix in perpetuity in the "sd-remove-unused-qubes". The rationale
is that this should be kept as long as possible, since it's possible
that some workstation may be "on the shelf" for some time, thus skipping
upadtes. By keeping its removal in the code, we can catch these
'straggler' machines and ensure proper removal.

(test on: openqa)