Make sd-proxy no longer internal
We now instruct users to restart and possibly run commands inside
sd-proxy to fix and debug their Tor connection, so we should make
it non-internal. This matches how sd-whonix was treated.

Note that the underlying sd-proxy-dvm is still internal, as users
should never explicitly interact with it.

Test this on OpenQA.