Simplify and improve testing of apt sources
For reasons that aren't super clear, the apt-get indextargets approach
has been flaky on both SDW CI and now OpenQA.

Simplify the test by just looking at the file we provision and not
the apt cache. AFAIK this part has always been relatively robust
so it shouldn't be flaky anymore.