Add note on why sd-remove-whonix... has cmd.run
'cmd.run' should be avoided but the native salt modules "pkg.del_repo"
and "pkg.purge" are likely getting confused by the fact that the debian
templates have both "dnf" and "apt" package managers. This happens in
Qubes because when Debian qubes are used as default (e.g. for
sys-firewall), they still need to enable dom0 to get updates (which are
through dnf).

Further diagnosis could be helpful, but these are temporary commands to
be used until Whonix 17 reaches EOL.

(test on: openqa)