Add PyQt6 but still install PyQt5 on python<=3.11
Poetry is used as a proxy for dom0 system dependencies, PyQt in
particular. This gives more flexibility on the development environment
(typically on Debian and not Fedora).

In this case, we wanted to target PyQt5 in Qubes 4.2, but PyQt6 in Qubes
4.3. The solution was to use the default python versions (3.11 and 3.13,
respectively) as proxies for said Qubes version. This is a simpler
approach and should work as long as the user sets the correct python
version via 'poetry env python3.<VERSION>.

Alternatives explored and why they were not chosen:

  - PyQt5 in 'dev' group and PyQt6 in optional 'dom0-fedora-41' group -
    this was considered, givent the fact that try..catch used for PtQt
    version imports first tries PyQt6. However, this would have required
    passing along the version information to dom0.

  - Separate dependecy groups 'dom0-fedora-37' (Qubes 4.2) and
    'dom0-fedora-41' (Qubes 4.3). This would have required further
    complicating the 'poetry install' command , with the need to specify
    the additional group with 'poetry install --with dom0-fedora-<VER>'.
    This would require properly documenting and changing all CI calls.