Work around salt module 'cmd.script' missing args
When running "sdw-admin --apply" "run-prep-upgrade-scripts" in
"sd-upgrade-templates.sls" and "sd-remove-unused-templates" as it fails
with the message:

  stdout: Please specify prepare or remove

This indicated that the 'args' parameter is not being properly
recognized. I looked at the saltstack documentation for the cmd module
and nothing indicates an 'args' parameters change. I've checked with
changelogs as well comparing the version in Qubes 4.2 with the one in
Qubes 4.3.

Instead of investigating how to get "cmd.script" this simply uses
another way to call the command as a temporary fix. For a more permanent
solution we need to fix this in all different places and if we choose to
go the 'cmd.run' route, we should place the script in some place other
than '/srv/salt/'.

Fixes #1499