Combine bootstrap-* into "make {dev|staging}"
Simplifies the deployment by consolidating these target. This way, a
developer only needs to remember / type one command to set up the
desired environment.

It also solves a bug where "make staging" was calling "scripts/prep-dev"
and thus always installing a locally built RPM (which goes against the
expectation of  the staging environment). The solution is use the
wrapper "install-rpm" which solves this nuance of where to fetch the RPM
from, depending on the environment.