Remove OpenQA NEEDLES_DIR override
The NEEDLES_DIR was a temporary override due to an issue in OpenQA [1],
where tests would fail with the following message:

    Gru job failed
    Reason: NOT updating dirty Git checkout at '/var/lib/openqa/share/tests/qubesos/needles'.
    In case this is expected (e.g. on a development openQA instance) you can disable auto-updating.
    Then the Git checkout will no longer be kept up-to-date, though. Checkout http://open.qa/docs/#_getting_tests for details.

This seems to have been resolved in the meantime, so it's time to
re-enable it. Another reason to do so is that with GUI tests we want to
be able to live-edit needles directly on OpenQA, interactively. But if
we override that directly it only knows about the ones in the repo and
not the ones we just created and therefore, it fails to find the needle.

Since at the time we had not yet actively worked on GUI test (which
need needles, we had not yet run into problems with our temporary
workaround)

[1]: https://github.com/QubesOS/openqa-tests-qubesos/pull/25#issuecomment-2965597793