Remove OpenQA NEEDLES_DIR override
The NEEDLES_DIR was a temporary override due to an experience issue in
OpenQA and now with GUI tests we're running into issues where it can't
find the needle we just created interactively. The assumption here is
that even though we do a live-creation of a needle, it's still fetching
the git-provided ones.

Since we had not yet actively worked on GUI test (which need needles, we
had not yet run into problems with our temporary workaround)