Suppress GPG prompt
Prompt was an unnecessary extra UX hoop as there was no reason for
denying the prompt (after all, the user always wants sd-app to be able
to have its submissions decrypted). The qrexec policy is a more
effective counter-measure, which is already in place: if it exists, the
GPG decryption is allowed. If it's not there, then it's not allowed.

Suppression is done by setting the max linux epoch value (2**31-1). This
works because When the qubes.Gpg RPC service is called, it compares the
expiry date with the current time.

Fixes #960

(test on: openqa)

[1]: https://github.com/QubesOS/qubes-app-linux-split-gpg/blob/bf8d6b38b58ad63bada2fce3c78a56aec2a421f6/qubes.Gpg.service#L32