sdw-admin: ignore applying to preloaded qubes
Preloaded qubes do not need to be configured as they automatically get
restarted when its template is changed and saved to disk. These
preloaded sd-viewer qubes were getting included for configuration due to
the inheritance of the 'sd-workstation' tag.

Fruthermore, this attempt at configuration was causing issues as
preloaded qubes would time-out when ordered to shut down (See #1475).

The implementation choice was to include this fix in configure() rather
than pre-filtering out preloaded qubes from the list of targets before
passing them on to the 'configure()' function. This was due to the fact
that the calling function should not need to be aware that preloaded
qubes do not (and should not) have configurations applied to them.

Fixes #1475

(test on: openqa)