Fix wrong unittest: 'X == Y' comparison behavior
Unittest' "setUp()" method somehow modifies class attributes set in it.
The end result is that when calling "assert self.x == "something" the
end comparison will be between "__repr__()" called on both ends. This is
undesireable behavior and fortunately when using pytest we got called
out to this problem. Further explanation can be found at [1].

[1]: https://github.com/freedomofpress/securedrop-workstation/issues/1338#issuecomment-3542447197