fix first-run make-dev via dom0 keyring
The `make dev` flow had an implicit dependency on the prod
`securedrop-workstation-keyring` package being already installed in dom0.
The dev and staging targets will also configure other repos, but in
all cases, we want the prod signing key and repo to be present,
otherwise the installation of `securedrop-workstation-dom0-config` will
fail.

Closes #1484.

Amended to squash in some review comments, and also to add:
test on: openqa