CI: use main's commit instead of 'main' git ref
The OpenQA CI job would fail when setting a commit status, because we
were calling with the provided git_ref on it ('main', in this case).
This obviously failed because 'main' is not a proper commit reference
[1].

This commit adds obtains this reference and uses it in the nightlies
job.

[1]: https://github.com/freedomofpress/securedrop-workstation/actions/runs/19700982727/job/56437179439#step:5:53