test: use pytest-xdist in dom0
Adds use of pytest-xdist to the dom0 integration test suite,
so that test execution can be parallelized. The updated Makefile target
uses automatic parallelization, which defaults to one thread per CPU
core. Notably, the batching of tests is set to be per-pytest-file,
to ensure that strictly ordered per-VM tests are still run in the proper
order. Performances gains on my Framework 13 significant, from ~5m to
~2.5m per invocation of "make test".

Also tacked on some timing information at the end of the test run, so
that developers can see the five slowest tests; those are currently the
MIME checks. We should be able to speed those up with parallelization
via params, in a subsequent PR.

Given the addition of new dom0 dependencies, I opted to bottle up the
deps in a makefile target, so it's easily callable in other contexts,
such as in the openqa test repo [0]. A corresponding PR must be made
into that repo in order for these changes to pass CI and be merged.

[0] https://github.com/freedomofpress/openqa-tests-qubesos