test: refactor per-VM dom0 modules
Reorganizes the dom0 integration tests:

  * prunes unused or duplicated Debian platform vars
  * renames tests to be clear about whether a specific VM is tested
  * moves per-VM tests into isolated files

Also updates the "check whether a service is running" logic to
deduplicate the tests.

Expands the "expected VMs" test with more descriptive var names,
for maintainability. Replaces the "qvm-kill" logic with "qvm-shutdown"
to treat the VMs kindly, in a nod to historical problems with killing,
such as [0].

Overall, the logic being tested remains the same: this change is mostly
to reorganize the tests so that subsequent, more substantive changes
(e.g. heavier use of pytest markers as mentioned in #1338) will have
more reviewable diffs.

[0] https://github.com/QubesOS/qubes-issues/issues/4890#issuecomment-606972698