Allow sd-gpg to obtain secret key in dom0 via qrexec
This implements a minimal Qubes RPC service that simply cats the secret
key file in dom0. Since it only cats a regular file and takes in no
parameters, it is security-wise. This serves as a replacement for
the salt logic that configure sd-gpg with the key file, effectively
clearing the path for a potentially disposable sd-gpg qube.

For context, all qubes can make certain qrexec calls to dom0 already
(see qubes.SyncAppMenus, for example).

Fixes #1505