test: use session-scoped tag vars
We were already running the "look up all SDW VMs by tag" logic in the
conftest.py, as a fixture, which enabled all tests to reuse the logic.
This change modifies the fixture logic to use session-based fixtures,
with autouse=True, which forces the VM lookup to run immediately on test
suite setup, rather than waiting for the first test to use the logic to
run, via the default setting of function-scoped fixtures [0].

The difference is that we can safely and reliably look up "all VMs
tagged as belonging to SDW" prior to running any tests, so that all
tests share the same view of the system state, notably *before* any
ephemeral DispVMs have been created.

In order to separate the launcher and dom0 integration test suites,
we provide per-suite `pytest.ini` config files, rather than unifying
the options under the parent project's `pyproject.toml`.

[0] https://docs.pytest.org/en/7.1.x/how-to/fixtures.html#fixture-scopes