Make sd-gpg load dom0-provided key on boot
Adds a new vm-config option called "run.<systemd-target>" to sd-gpg.
The idea is that some component on the client repo will read this and
execute it before said systemd target.

The initial use-case implemented is to import the gpg key from dom0 into
sd-gpg. Compared to having a dedicated component in sd-gpg is with this
this approach allows for making changes that apply on boot in said qube,
but without needing to perform changes in packages inside the VM itself.

This was inspired by a discussion on the Qubes forum:
https://forum.qubes-os.org/t/customize-a-named-disposable-using-the-vm-config-feature/34207